+++ src/index.mjs
@@ -1,8 +1,2 @@
-/*!
- * https://github.com/jaydenseric/graphql-react
- * @author Jayden Seric
- * @license MIT
- */
-
 export * from './graphql'
 export * from './components'